# goserver/api/openapi.yaml
openapi: 3.0.0
info:
  title: WebRTC Signaling Server
  description: Python WebRTC signaling server with WebSocket support
  version: 1.0.0
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      description: |
        Create a new user account with password, and username.

        **Request body:**
        ```json
        {
          "password": "string",
          "name": "John Doe"
        }
        ```

        **Successful response:**
        Returns `access_token` and `refresh_token` with expiration times.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
              examples:
                success:
                  summary: Successful registration
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refresh_token: "dGhpc2lzYXJlZnJlc2h0b2tlbg..."
                    expires_in: 3600
                    refresh_expires_in: 86400
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictUserError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: User login
      description: |
        Authenticate user with username and password.

        **Request body:**
        ```json
        {
          "username": "usertest",
          "password": "string"
        }
        ```

        **Successful response:**
        Returns `access_token` and `refresh_token` with expiration times.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
              examples:
                success:
                  summary: Successful login
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refresh_token: "dGhpc2lzYXJlZnJlc2h0b2tlbg..."
                    expires_in: 3600
                    refresh_expires_in: 86400
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Exchange a valid `refresh_token` for a new `access_token`.

        **Request body:**
        ```json
        {
          "refresh_token": "string"
        }
        ```

        **Successful response:**
        Returns a new `access_token` with expiration time.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Successful token refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
              examples:
                success:
                  summary: Successful token refresh
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refresh_token: "dGhpc2lzYXJlZnJlc2h0b2tlbg..."
                    expires_in: 3600
                    refresh_expires_in: 86400
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      description: |
        Invalidate the current session tokens.

        Requires a valid `Authorization: Bearer <access_token>` header.
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Successfully logged out (no content returned)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/me:
    get:
      tags: [User]
      summary: Get current user profile
      description: |
        Retrieve details of the authenticated user.

        Requires a valid `Authorization: Bearer <access_token>` header.
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              examples:
                success:
                  summary: Current user profile
                  value:
                    id: "123"
                    name: "John Doe"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/users:
    get:
      tags: [User]
      summary: List all registered users on the server
      description: |
        Returns every user account (directory / address book). Optional `avatar_url` is present when the user has uploaded an avatar.

        Requires a valid `Authorization: Bearer <access_token>` header.
      operationId: listServerUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/rooms:
    post:
      tags: [Signaling]
      summary: Create a new signaling room
      description: Create a new room that clients can join for WebRTC peer-to-peer calls.
      operationId: createRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          $ref: '#/components/responses/ServerError'
    get:
      tags: [Signaling]
      summary: List available rooms
      operationId: listRooms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of available rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/rooms/{roomId}/join:
    post:
      tags: [Signaling]
      summary: Join a signaling room
      description: Join an existing room. The WebSocket endpoint is used after joining.
      operationId: joinRoom
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully joined room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomJoinResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/rooms/{roomId}/participants:
    get:
      tags: [Signaling]
      summary: Get room participants
      description: Get a list of all users currently connected to the room.
      operationId: getRoomParticipants
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of room participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomParticipantsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/chat/ws:
    get:
      tags: [Chat]
      summary: WebSocket connection for text chat
      description: |
        Upgrade to a WebSocket connection for real-time text chat in a room.
        Authenticate via `token` query param (JWT). Room is specified via `room` query param.
      operationId: chatWebSocket
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: room
          in: query
          required: true
          schema:
            type: string
        - name: username
          in: query
          required: false
          schema:
            type: string
      responses:
        '101':
          description: Switching Protocols (WebSocket upgrade)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [password, username]
      properties:
        password:
          type: string
          format: password
          minLength: 6
        username:
          type: string
          minLength: 3
          maxLength: 20
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Access token expiration time in seconds
        refresh_expires_in:
          type: integer
          description: Refresh token expiration time in seconds
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar_url:
          type: string
          description: Present when the user has an avatar; relative path e.g. /api/v1/avatars/{username}
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          enum:
            - INVALID_CREDENTIALS
            - TOKEN_EXPIRED
            - INVALID_REFRESH_TOKEN
            - UNAUTHORIZED
            - VALIDATION_ERROR
            - NOT_FOUND
            - TOO_MANY_REQUESTS
            - SERVER_ERROR
        message:
          type: string
      required: [code, message]
    CreateRoomRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Human-readable room name
        type:
          type: string
          enum: [voice, text]
          default: voice
          description: Channel type
    Room:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [voice, text]
          description: Channel type
        created_at:
          type: string
          format: date-time
        roommates:
          type: array
          items:
            type: string
            description: A list of participatns of room
    RoomJoinResponse:
      type: object
      properties:
        room_id:
          type: string
        client_id:
          type: string
          description: Unique ID assigned to this client in the room
    RoomParticipantsResponse:
      type: object
      properties:
        room_id:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/RoomParticipant'
    RoomParticipant:
      type: object
      properties:
        client_id:
          type: string
          description: Unique client ID in the room
        username:
          type: string
          description: Username of the participant
...
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            badRequest:
              value:
                code: "VALIDATION_ERROR"
                message: "Invalid input data."
    ConflictUserError:
      description: Conflict - User already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              value:
                code: "USER_EXISTS"
                message: "User with this username already exists"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: "UNAUTHORIZED"
                message: "Missing or invalid access token."
    UnprocessableEntityError:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unprocessable:
              value:
                code: "VALIDATION_ERROR"
                message: "Unable to process the request due to semantic errors."
    NotFoundError:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: "NOT_FOUND"
                message: "The requested resource does not exist."
...
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              value:
                code: "SERVER_ERROR"
                message: "An unexpected error occurred."