# goserver/internal/gen/api/http.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for CreateRoomRequestType.
const (
	CreateRoomRequestTypeText  CreateRoomRequestType = "text"
	CreateRoomRequestTypeVoice CreateRoomRequestType = "voice"
)

// Defines values for ErrorResponseCode.
const (
	INVALIDCREDENTIALS  ErrorResponseCode = "INVALID_CREDENTIALS"
	INVALIDREFRESHTOKEN ErrorResponseCode = "INVALID_REFRESH_TOKEN"
	NOTFOUND            ErrorResponseCode = "NOT_FOUND"
	SERVERERROR         ErrorResponseCode = "SERVER_ERROR"
	TOKENEXPIRED        ErrorResponseCode = "TOKEN_EXPIRED"
	TOOMANYREQUESTS     ErrorResponseCode = "TOO_MANY_REQUESTS"
	UNAUTHORIZED        ErrorResponseCode = "UNAUTHORIZED"
	VALIDATIONERROR     ErrorResponseCode = "VALIDATION_ERROR"
)

// Defines values for RoomType.
const (
	RoomTypeText  RoomType = "text"
	RoomTypeVoice RoomType = "voice"
)

// AuthTokens defines model for AuthTokens.
type AuthTokens struct {
	AccessToken *string `json:"access_token,omitempty"`

	// ExpiresIn Access token expiration time in seconds
	ExpiresIn *int `json:"expires_in,omitempty"`

	// RefreshExpiresIn Refresh token expiration time in seconds
	RefreshExpiresIn *int    `json:"refresh_expires_in,omitempty"`
	RefreshToken     *string `json:"refresh_token,omitempty"`
}

// CreateRoomRequest defines model for CreateRoomRequest.
type CreateRoomRequest struct {
	// Name Human-readable room name
	Name string `json:"name"`

	// Type Channel type
	Type *CreateRoomRequestType `json:"type,omitempty"`
}

// CreateRoomRequestType Channel type
type CreateRoomRequestType string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    ErrorResponseCode `json:"code"`
	Message string            `json:"message"`
}

// ErrorResponseCode defines model for ErrorResponse.Code.
type ErrorResponseCode string

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// RefreshRequest defines model for RefreshRequest.
type RefreshRequest struct {
	RefreshToken string `json:"refresh_token"`
}

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// Room defines model for Room.
type Room struct {
	CreatedAt *time.Time `json:"created_at,omitempty"`
	Id        *string    `json:"id,omitempty"`
	Name      *string    `json:"name,omitempty"`
	Roommates *[]string  `json:"roommates,omitempty"`

	// Type Channel type
	Type *RoomType `json:"type,omitempty"`
}

// RoomType Channel type
type RoomType string

// RoomJoinResponse defines model for RoomJoinResponse.
type RoomJoinResponse struct {
	// ClientId Unique ID assigned to this client in the room
	ClientId *string `json:"client_id,omitempty"`
	RoomId   *string `json:"room_id,omitempty"`
}

// RoomParticipant defines model for RoomParticipant.
type RoomParticipant struct {
	// ClientId Unique client ID in the room
	ClientId *string `json:"client_id,omitempty"`

	// Username Username of the participant
	Username *string `json:"username,omitempty"`
}

// RoomParticipantsResponse defines model for RoomParticipantsResponse.
type RoomParticipantsResponse struct {
	Participants *[]RoomParticipant `json:"participants,omitempty"`
	RoomId       *string            `json:"room_id,omitempty"`
}

// UserProfile defines model for UserProfile.
type UserProfile struct {
	// AvatarUrl Present when the user has an avatar; relative path e.g. /api/v1/avatars/{username}
	AvatarUrl *string `json:"avatar_url,omitempty"`
	Id        *string `json:"id,omitempty"`
	Username  *string `json:"username,omitempty"`
}

// BadRequestError defines model for BadRequestError.
type BadRequestError = ErrorResponse

// ConflictUserError defines model for ConflictUserError.
type ConflictUserError = ErrorResponse

// NotFoundError defines model for NotFoundError.
type NotFoundError = ErrorResponse

// ServerError defines model for ServerError.
type ServerError = ErrorResponse

// UnauthorizedError defines model for UnauthorizedError.
type UnauthorizedError = ErrorResponse

// UnprocessableEntityError defines model for UnprocessableEntityError.
type UnprocessableEntityError = ErrorResponse

// ChatWebSocketParams defines parameters for ChatWebSocket.
type ChatWebSocketParams struct {
	Token    string  `form:"token" json:"token"`
	Room     string  `form:"room" json:"room"`
	Username *string `form:"username,omitempty" json:"username,omitempty"`
}

// LoginUserJSONRequestBody defines body for LoginUser for application/json ContentType.
type LoginUserJSONRequestBody = LoginRequest

// RefreshTokenJSONRequestBody defines body for RefreshToken for application/json ContentType.
type RefreshTokenJSONRequestBody = RefreshRequest

// RegisterUserJSONRequestBody defines body for RegisterUser for application/json ContentType.
type RegisterUserJSONRequestBody = RegisterRequest

// CreateRoomJSONRequestBody defines body for CreateRoom for application/json ContentType.
type CreateRoomJSONRequestBody = CreateRoomRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// User login
	// (POST /api/v1/auth/login)
	LoginUser(w http.ResponseWriter, r *http.Request)
	// Logout user
	// (POST /api/v1/auth/logout)
	LogoutUser(w http.ResponseWriter, r *http.Request)
	// Refresh access token
	// (POST /api/v1/auth/refresh)
	RefreshToken(w http.ResponseWriter, r *http.Request)
	// Register new user
	// (POST /api/v1/auth/register)
	RegisterUser(w http.ResponseWriter, r *http.Request)
	// WebSocket connection for text chat
	// (GET /api/v1/chat/ws)
	ChatWebSocket(w http.ResponseWriter, r *http.Request, params ChatWebSocketParams)
	// Get current user profile
	// (GET /api/v1/me)
	GetCurrentUser(w http.ResponseWriter, r *http.Request)
	// List all registered users on the server
	// (GET /api/v1/users)
	ListServerUsers(w http.ResponseWriter, r *http.Request)
	// List available rooms
	// (GET /api/v1/rooms)
	ListRooms(w http.ResponseWriter, r *http.Request)
	// Create a new signaling room
	// (POST /api/v1/rooms)
	CreateRoom(w http.ResponseWriter, r *http.Request)
	// Join a signaling room
	// (POST /api/v1/rooms/{roomId}/join)
	JoinRoom(w http.ResponseWriter, r *http.Request, roomId string)
	// Get room participants
	// (GET /api/v1/rooms/{roomId}/participants)
	GetRoomParticipants(w http.ResponseWriter, r *http.Request, roomId string)
	// WebSocket connection for signaling
	// (GET /api/v1/ws)
	SignalingWebSocket(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// LoginUser operation middleware
func (siw *ServerInterfaceWrapper) LoginUser(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.LoginUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// LogoutUser operation middleware
func (siw *ServerInterfaceWrapper) LogoutUser(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.LogoutUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RefreshToken operation middleware
func (siw *ServerInterfaceWrapper) RefreshToken(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RefreshToken(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RegisterUser operation middleware
func (siw *ServerInterfaceWrapper) RegisterUser(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RegisterUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ChatWebSocket operation middleware
func (siw *ServerInterfaceWrapper) ChatWebSocket(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ChatWebSocketParams

	// ------------- Required query parameter "token" -------------

	if paramValue := r.URL.Query().Get("token"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "token"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "token", r.URL.Query(), &params.Token)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "token", Err: err})
		return
	}

	// ------------- Required query parameter "room" -------------

	if paramValue := r.URL.Query().Get("room"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "room"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "room", r.URL.Query(), &params.Room)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "room", Err: err})
		return
	}

	// ------------- Optional query parameter "username" -------------

	err = runtime.BindQueryParameter("form", true, false, "username", r.URL.Query(), &params.Username)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "username", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ChatWebSocket(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetCurrentUser operation middleware
func (siw *ServerInterfaceWrapper) GetCurrentUser(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetCurrentUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListServerUsers operation middleware
func (siw *ServerInterfaceWrapper) ListServerUsers(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListServerUsers(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListRooms operation middleware
func (siw *ServerInterfaceWrapper) ListRooms(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListRooms(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateRoom operation middleware
func (siw *ServerInterfaceWrapper) CreateRoom(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateRoom(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// JoinRoom operation middleware
func (siw *ServerInterfaceWrapper) JoinRoom(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "roomId" -------------
	var roomId string

	err = runtime.BindStyledParameterWithOptions("simple", "roomId", r.PathValue("roomId"), &roomId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "roomId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.JoinRoom(w, r, roomId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRoomParticipants operation middleware
func (siw *ServerInterfaceWrapper) GetRoomParticipants(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "roomId" -------------
	var roomId string

	err = runtime.BindStyledParameterWithOptions("simple", "roomId", r.PathValue("roomId"), &roomId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "roomId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRoomParticipants(w, r, roomId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SignalingWebSocket operation middleware
func (siw *ServerInterfaceWrapper) SignalingWebSocket(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SignalingWebSocket(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("POST "+options.BaseURL+"/api/v1/auth/login", wrapper.LoginUser)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/auth/logout", wrapper.LogoutUser)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/auth/refresh", wrapper.RefreshToken)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/auth/register", wrapper.RegisterUser)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/chat/ws", wrapper.ChatWebSocket)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/me", wrapper.GetCurrentUser)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/users", wrapper.ListServerUsers)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/rooms", wrapper.ListRooms)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/rooms", wrapper.CreateRoom)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/rooms/{roomId}/join", wrapper.JoinRoom)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/rooms/{roomId}/participants", wrapper.GetRoomParticipants)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/ws", wrapper.SignalingWebSocket)

	return m
}