# vueclient/src/api/models/AuthTokens.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type AuthTokens = {
    access_token?: string;
    refresh_token?: string;
    /**
     * Access token expiration time in seconds
     */
    expires_in?: number;
    /**
     * Refresh token expiration time in seconds
     */
    refresh_expires_in?: number;
};