# vueclient/src/api/models/CreateRoomRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type CreateRoomRequest = {
    /**
     * Human-readable room name
     */
    name: string;
    /**
     * Channel type
     */
    type?: 'voice' | 'text';
};