# vueclient/src/api/models/RoomJoinResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type RoomJoinResponse = {
    room_id?: string;
    /**
     * Unique ID assigned to this client in the room
     */
    client_id?: string;
};