# vueclient/src/api/models/RoomParticipantsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
import type { RoomParticipant } from './RoomParticipant';
export type RoomParticipantsResponse = {
    room_id?: string;
    participants?: Array<RoomParticipant>;
};